refactor: add error types with thiserror and anyhow
- Add thiserror, anyhow, tracing, async-trait dependencies
- Add Cargo.toml metadata (authors, description, license, repository)
- Add release profile with strip and LTO
- Create AppError enum in src/error.rs
- Switch main() to return anyhow::Result
- Fix all pre-existing clippy warnings (manual_strip, lines_filter_map_ok,
  unnecessary_cast, manual_div_ceil, collapsible_if)

+++ src/system.rs
@@ -249,7 +249,7 @@ impl SystemReader {
                 if let Some(Ok(bat)) = batteries.next() {
                     let pct = bat.state_of_charge().value * 100.0;
                     let status = format!("{:?}", bat.state());
-                    (pct as f32, status)
+                    (pct, status)
                 } else {
                     (self.sim.battery_percent, "Unknown".to_string())
                 }
